code-driven: rename air-quality-server cpp to AirQualityCluster
Rename air-quality-server.cpp to AirQualityCluster.cpp as the first step
in migrating the Air Quality cluster to a code-driven implementation.
Header file is kept as-is for backward compatibility with code generation.

rename air-quality-server.h to AirQualityCluster.h
