+++ src/app/clusters/air-quality-server/AirQualityCluster.cpp
@@ -19,7 +19,7 @@
 #include "app-common/zap-generated/ids/Clusters.h"
 #include <app-common/zap-generated/attributes/Accessors.h>
 #include <app/AttributeAccessInterfaceRegistry.h>
-#include <app/clusters/air-quality-server/air-quality-server.h>
+#include <app/clusters/air-quality-server/AirQualityCluster.h>
 #include <app/reporting/reporting.h>
 #include <app/util/attribute-storage.h>
