+++ src/app/clusters/air-quality-server/app_config_dependent_sources.gni
@@ -12,6 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 app_config_dependent_sources = [
-  "air-quality-server.cpp",
+  "AirQualityCluster.cpp",
   "air-quality-server.h",
 ]
